+++ lib/src/brazil/cnpj.dart
@@ -3,15 +3,18 @@ import 'cnpj_strip.dart';
 
 /// CNPJ - Cadastro Nacional de Pessoa Jurídica.
 ///
-/// Kind of brazilian Company's Registered Number or National Registry of Legal
+/// Kind of Brazilian Company's Registered Number or National Registry of Legal
 /// Entities.
+///
+/// If CNPJ is mandatory, see [ReqCnpj] or [Req].
 class Cnpj {
-  /// Validates an optional cnpj field.
+  /// Validates an optional CNPJ field.
   ///
-  /// [mal] is the error message in case of a malformed CNPJ; defaults to
+  /// [mal] the error message in case of a malformed CNPJ; the default value is
   /// 'malformed cnpj'.
   const Cnpj({String? mal}) : _malformed = mal ?? 'malformed cnpj';
 
+  /// The error message.
   final String _malformed;
 
   /// Returns null if [cnpj] is well-formed; otherwise, the erro message.