+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 ## [Unreleased]
 
+### Added
+
+- 'ReqCnpj' which is the composition between the validators 'Req' and 'Cnpj' —
+  [65](https://github.com/dartoos-dev/formdator/issues/65).
+- 'ReqCpf' which is the composition between the validators 'Req' and 'Cpf' —
+  [65](https://github.com/dartoos-dev/formdator/issues/66).
+
 ## [0.7.1] - 2021-07-21
 
 ### Added